requirements.txt
fastapi
uvicorn
python-multipart
ollama
pika
oracledb
gradio
psycopg2-binary
redis
requests
pandas
python-dotenv
gitpython
gradio[mcp]
mermaid-py
ipython
sqlparse
gradio
sqlcompyre
sqlalchemy
cx_Oracle
pypdf
langchain
pgvector
langchain_community 
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
langchain_ollama
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-psycopg2
